move seed client def
