+++ packages/seed-helpers/src/index.ts
@@ -17,7 +17,7 @@
 export {
   createSeedWithMetadata,
   SeedBuilder,
-  type SeedClient,
   type SeedFunction,
 } from "./SeedBuilder.js";
 export type { SeedLinkEntry, SeedOutput, SeedProps, SeedRef } from "./types.js";
+export type { LinkTargetType, LinkTargets } from "./linkTypes.js";
